Add local cache support for downloads.

+++ vscode/tools/update/update-sd-card.sh
@@ -22,11 +22,13 @@ fatal() {
 usage() {
 	local filename
 	filename=$(basename "$0")
-	log "Usage: $filename [-f <format-mmc.sh>] <sd-card> [<zuul-uuid>] [<rootfs>]"
+	log "Usage: $filename [-c] [-f <format-mmc.sh>] <sd-card> [<zuul-uuid>] [<rootfs>]"
 	log "where:"
-	log "    sd-card     sd card device /dev/XXX"
-	log "    zuul-uuid   zuul periodic build pipeline identifier (UUID)"
-	log "    rootfs      rootfs file name"
+	log "    sd-card             sd card device /dev/XXX"
+	log "    zuul-uuid           zuul periodic build pipeline identifier (UUID)"
+	log "    rootfs              rootfs file name"
+	log "    -f <format-mmc.sh>  name of the script should be executed to flash MMC"
+	log "    -c                  clean any cached data, force download"
 	log ""
 	log "examples:"
 	log "    $filename sda 6cba2361dc4f4e258dc258b425828f4a"
@@ -56,20 +58,24 @@ run() {
 }
 
 download_file() {
-	logn "Downloading $1... "
 	local target
 	target=$(basename "$1")
-	if ! run safe wget --quiet "$zuul_source/$1" -O "$download_dir/$target"; then
-		log "FAILED"
-		fatal "Faild to download '$zuul_source/$1'."
+	download_name="$download_dir/$target.$arg_zuul_uuid"
+	if [[ -f "$download_name" ]]; then
+		log "Using $1 from cache... OK"
+		return
 	fi
-	log "OK"
+	if ! wget -q --show-progress "$zuul_source/$1" -O "$download_name.dld"; then
+		fatal "Failed to download '$zuul_source/$1'."
+	fi
+	mv "$download_name.dld" "$download_name"
 }
 
 arg_sd_device=""
 arg_zuul_uuid=""
 arg_rootfs_squashfs_path=""
 arg_format_mmc_path=""
+arg_purge_cache=""
 
 while [[ "$#" != "0" ]]; do
 	case "$1" in
@@ -80,6 +86,10 @@ while [[ "$#" != "0" ]]; do
 		if [[ ! -f "$arg_format_mmc_path" ]]; then
 			fatal "Argument '-f': unable to find file '$arg_format_mmc_path'."
 		fi
+		;;
+	-c)
+		arg_purge_cache="1"
+		shift
 		;;
 	*)
 		break
@@ -121,17 +131,19 @@ format_mmc_name="format-mmc.sh"
 rootfs_squashfs_name="rootfs.squashfs"
 
 run mkdir -p "$download_dir"
-run rm -rf "${download_dir:?}/"*
+if [[ "$arg_purge_cache" != "" ]]; then
+	run rm -rf "${download_dir:?}/"*
+fi
 
 if [[ "$arg_zuul_uuid" != "" ]]; then
 	if [[ "$arg_format_mmc_path" == "" ]]; then
 		download_file "$format_mmc_name"
-		arg_format_mmc_path="$download_dir/$format_mmc_name"
+		arg_format_mmc_path="$download_name"
 		run chmod +x "$arg_format_mmc_path"
 	fi
 	if [[ "$arg_rootfs_squashfs_path" = "" ]]; then
 		download_file "$rootfs_squashfs_name"
-		arg_rootfs_squashfs_path="$download_dir/$rootfs_squashfs_name"
+		arg_rootfs_squashfs_path="$download_name"
 	fi
 fi
